ring 3a fix: cast ring output back to input dtype (bf16) for o-proj matmul
Run sd2qh (57828b3): KERNEL COMPILED (past all NKI compile errors) but NEFF exec setup
failed: 'matmul: input datatypes mismatched' at Linear[o-proj]. Root cause: return_partials
makes the kernel result fp32 (needed for cross-rank combine precision), and _attend_ring's
acc/l stays fp32; that fp32 flows into self.o (bf16 weights) -> dtype mismatch. Default path
returns q.dtype(bf16). Fix: out = out.to(q_kern.dtype) before returning — combine stays fp32
internally, output matches the default bf16 contract.

+++ models/dit_attention.py
@@ -483,7 +483,11 @@ class CausalWanSelfAttention(nn.Module):
                 l = l * cp + sum_s * cc
                 acc = acc * cp + O_s * cc
                 m = m_new
-        out = acc / l                              # [Sq, bs, d]
+        out = acc / l                              # [Sq, bs, d] fp32
+        # combine runs in fp32 (kernel partials are fp32 for cross-rank precision), but the
+        # default kernel returns q.dtype (bf16) and the downstream Linear (self.o) expects
+        # that — cast back or the o-proj matmul hits 'input datatypes mismatched'.
+        out = out.to(q_kern.dtype)
         # match default _attend contract exactly: out.unsqueeze(0).flatten(2)
         # -> [1, Sq, bs*d]  (same as wan_flash_self_attn result path).
         return out.unsqueeze(0).flatten(2)
